Android fundamentals 01.2 Part B: The layout editor Homework
