+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdkVersion 31
 
     defaultConfig {
-        applicationId "com.example.hellotoast"
+        applicationId "com.example.helloconstraint"
         minSdkVersion 21
         targetSdkVersion 31
         versionCode 1
