+++ app/src/androidTest/java/com/example/helloconstraint/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.hellotoast
+package com.example.helloconstraint
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
