--- app/src/main/java/com/example/hellotoast/MainActivity.kt
@@ -1,29 +1,0 @@
-package com.example.hellotoast
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-import android.widget.TextView
-import android.widget.Toast
-
-class MainActivity : AppCompatActivity() {
-
-    private var mCount: Int = 0
-    private var mShowCount: TextView? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        mShowCount = findViewById(R.id.show_count)
-    }
-
-    fun countUp(view: View) {
-        mCount++
-        mShowCount?.text = mCount.toString()
-    }
-
-    fun showToast(view: View) {
-        val toast = Toast.makeText(this, R.string.toast_message, Toast.LENGTH_SHORT)
-        toast.show()
-    }
-}
