+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,8 @@
 <resources>
-    <string name="app_name">Hello Toast</string>
+    <string name="app_name">Hello Constraint</string>
     <string name="button_label_count">Count</string>
     <string name="button_label_toast">Toast</string>
     <string name="count_initial_value">0</string>
     <string name="toast_message">Hello Toast!</string>
+    <string name="button_zero">0</string>
 </resources>
